Move ResourceSegmentsConverter into shardingsphere-proxy-backend module
